test,refactor: move rfactor tests
